Extract DuckDB queries
